Problem 33: There are 4 companies in the markets of Chittagong, A, B, C, and D. This year,
Company A made a profit of P lakh,

Company B made a profit of Q lakh,

Company C made a profit of R lakh,

Company D made a profit of S lakh.

There is said to be a monopoly in the market if the profit made by one company is strictly
greater than the sum of profits made by all other companies. Determine if there is a monopoly
in the market or not.